pwn_Snowstorm/checker/requirements.txt
pwntools==4.13.1